moves interfaces to a seperate file and import where needed
